+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.4.3
 
 - added correct support of condition result to opcodes 0AB1 0AB2. Fixes possible bugs when 0AB1 are used in multi conditional if statements. 
+- set condition result in 0ADA and 0AD8
 
 ## 4.4.2